+++ app/frontend/websocket/handlers/__tests__/symbol-trigger-sell.test.js
@@ -36,6 +36,7 @@ describe('symbol-trigger-sell.test.js', () => {
     beforeEach(async () => {
       const { logger } = require('../../../../helpers');
       mockLogger = logger;
+      mockLogger.fields = { correlationId: 'correlationId' };
 
       const { handleSymbolTriggerSell } = require('../symbol-trigger-sell');
       await handleSymbolTriggerSell(mockLogger, mockWebSocketServer, {
@@ -59,7 +60,9 @@ describe('symbol-trigger-sell.test.js', () => {
     });
 
     it('triggers queue.executeFor', () => {
-      expect(mockQueue.executeFor).toHaveBeenCalledWith(mockLogger, 'BTCUSDT');
+      expect(mockQueue.executeFor).toHaveBeenCalledWith(mockLogger, 'BTCUSDT', {
+        correlationId: 'correlationId'
+      });
     });
 
     it('triggers ws.send', () => {